messages/protobuf: Make typedMessageFromPb common utility function
This function will help to convert any message from Protobuf into API
representation.
