+++ messages/protobuf/impl.go
@@ -69,25 +69,6 @@ func (*impl) NewViewChange(r uint32, nv uint64, log messages.MessageLog, vcCert 
 	panic("Not implemented")
 }
 
-func typedMessageFromPb(pbMsg *pb.Message) (messages.Message, error) {
-	switch t := pbMsg.Typed.(type) {
-	case *pb.Message_Hello:
-		return newHelloFromPb(t.Hello)
-	case *pb.Message_Request:
-		return newRequestFromPb(t.Request)
-	case *pb.Message_Reply:
-		return newReplyFromPb(t.Reply)
-	case *pb.Message_Prepare:
-		return newPrepareFromPb(t.Prepare)
-	case *pb.Message_Commit:
-		return newCommitFromPb(t.Commit)
-	case *pb.Message_ReqViewChange:
-		return newReqViewChangeFromPb(t.ReqViewChange)
-	default:
-		return nil, xerrors.New("unknown message type")
-	}
-}
-
 func marshalMessage(m proto.Message) ([]byte, error) {
 	return proto.Marshal(pb.WrapMessage(m))
 }
